+++ spotify_ripper/utils.py
@@ -1,3 +1,5 @@
+# -*- coding: utf8 -*-
+
 from __future__ import unicode_literals
 
 import os, sys
